// src/typings/api/trade.d.ts
export type ICreateFutureNormalOrderStrategy = {
  /**
   * 触发价格
   */
  triggerPrice: string
  /**
   * 触发价格类型（mark 标记，new 最新）
   */
  triggerPriceTypeInd: string
  /**
   * 方向 close_long 平多 , close_short 平空
   */
  triggerSideInd: string
  /**
   * 委托价格类型 limit 限价 market 市价
   */
  entrustTypeInd: string
  /**
   * 委托价格，委托价格类型未 market 不传；limit 必传
   */
  price?: string
  /**
   * 委托数量 market limit 都必传
   */
  size?: string
  /** 触发方向 */
  triggerDirectionInd: string
  strategyTypeInd: string
}

export type OrderDragArrayType = {
  sort: number
  value: string
  name: string
}

export type ICreateFutureNormalOrderReq = {
  /**
   * 合约组 id
   */
  groupId: any
  /**
   * 仓位 id ,减仓单，平仓单必传
   */
  positionId: number
  /**
   * 交易对 id
   */
  tradeId: number
  /**
   * 订单类型    limit_order 限价委托单  market_order 市价委托单 forced_liquidation_order 强平委托单 forced_lighten_order 强减委托单
   */
  typeInd: string
  /**
   * 委托价格类型 limit 限价 market 市价
   */
  entrustTypeInd: string
  /**
   * 方向类型 open_long 开多 , open_short 开空 ,close_long 平多，close_short 平空
   */
  sideInd: string
  /**
   * 委托价格;限价单必传，市价单不传
   */
  price?: number
  /**
   * 委托数量;限价单市价单必传
   */
  size?: number
  /**
   * 市价单单位 amount 按金额 quantity 按数量
   */
  marketUnit?: string
  /**
   * 下单金额或减仓价值，市价单按金额 market_unit=amount 时必传
   */
  funds?: number
  /**
   * 开仓保证金来源  assets 使用资产账户的资金开仓 cgroup 使用当前合约组的额外保证金开仓
   */
  marginType?: string
  /**
   * 开仓初始仓位保证金
   */
  initMargin?: number
  /**
   * 开仓额外保证金来源  assets 账户资产作为额外保证金 open_funds 开仓资金作为额外保证金
   */
  additionalMarginType?: string
  /**
   * 开仓额外保证金
   */
  additionalMargin?: number
  /**
   * 开仓杠杆倍数
   */
  lever?: number
  placeUnit: string
  autoAddMargin?: string
  strategy?: {
    stopProfit?: ICreateFutureNormalOrderStrategy
    stopLoss?: ICreateFutureNormalOrderStrategy
  }
  accountType?: string
  coupons?: {
    couponId: number
    couponType: string
    couponAmount: number
  }[]
  voucherAmount?:number
}
export type ICreateFuturePlanOrderReq = YapiPostV1PerpetualPlanOrdersPlaceApiRequest
// ...
 export type IFutureCoinDetail = YapiGetV1PerpetualTradePairDetailData & {
  assetFeeRate: number
 }

 export type ITradePairLever = Required<YapiGetV1PerpetualTradePairDetailData>['tradePairLeverList'][0]